students/k3340/Zaharkin_Bogdan/Lr2/Task2/requirements.txt
sqlmodel
asyncpg
python-dotenv
aiohttp
bs4
requests
psycopg2-binary
chardet
